backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
pandas
scikit-learn
joblib
torch
torchvision
sentence-transformers
ftfy
regex
tqdm